requirements.txt
jsonschema==4.23.0
pyyaml==6.0.2
networkx==3.3
rich==13.9.4
pytest==8.3.2
ruff==0.6.9
